Module PredicateAutomata.MakeBounded

Parameters

A : S

Signature

val bounded_empty : A.t -> int -> (A.letter * int) list option

bounded_empty pa bound finds Some word which is accepted by pa and uses only indexed letters with index <= bound; None if there are no such words.

val bounded_invariant : A.t -> int -> A.formula -> (A.letter * int) list option