Module PredicateAutomata.MakeEmpty

Parameters

A : sig ... end

Signature

type solver

val pp : Stdlib.Format.formatter -> solver -> unit

val mk_solver : A.t -> solver

val conjoin_transition : solver -> A.predicate -> A.letter_set -> A.formula -> unit

conjoin_transition solver q sigma phi adds destructively updates the PA in solver by adding the transition rule delta(q,sigma) = phi /\ psi, where psi was the previous transition rule for (q,sigma).

val add_predicate : solver -> A.predicate -> int -> unit

Add a predicate to the vocabulary of a solver. Initially, delta(q,sigma) = true for all letters sigma.

val add_accepting_predicate : solver -> A.predicate -> int -> unit

Add a predicate to the vocabulary of a solver, and make it accepting. Initially, delta(q,sigma) = true for all letters sigma.

val mem_vocabulary : solver -> A.predicate -> bool

val find_word : ?⁠max_index:int -> solver -> (A.letter * int) list option

val alphabet : solver -> A.letter_set

val vocabulary : solver -> (A.predicate * int) BatEnum.t