Module S.Config

type t = config

type predicate = predicate

module Predicate : Struct.Symbol with type t = predicate

val pp : Stdlib.Format.formatter -> t -> unit

val show : t -> string

val equal : t -> t -> bool

val hash : t -> int

val compare : t -> t -> int

val props : t -> (predicate * int list) BatEnum.t

Enumerate the propositions which hold in a structure

val preds : t -> predicate BatEnum.t

val universe_size : t -> int

val universe : t -> int BatEnum.t

val make : ?⁠size:int -> (predicate * int list) BatEnum.t -> t

val models : ?⁠env:int list -> t -> (predicate, int) PaFormula.formula -> bool

val min_models : ?⁠env:int list -> int -> (predicate, int) PaFormula.formula -> t BatEnum.t

Find all minimal models of a given formula, with a given universe

val substructure : t -> t -> bool

Is the first structure a substructure of the second?

val embeds : t -> t -> bool

Is there an embedding (injective homomorphism) from the first structure into the second?

val union : t -> t -> t

union s s' is a structure whose universe is the union of the universes of s and s' and whose interpretation of each predicate p is the union of its interpretation in s and s'.

val empty : int -> t

val full : (predicate * int) BatEnum.t -> int -> t

full vocabulary is the largest structure over the given vocabulary

val add : predicate -> int list -> t -> t

val remove : predicate -> int list -> t -> t

val minimize : ?⁠env:int list -> t -> (predicate, int) PaFormula.formula -> t

Given a structure m and positive formula phi such that m |= phi, find a minimal substructure m' of m such that m' |= phi

val instantiate : ?⁠env:int list -> t -> (predicate, int) PaFormula.formula -> (predicate, int) PaFormula.formula option

Given a structure m and positive formula phi, if m |= phi then strengthen phi by removing disjuncts and instantiating existential quantifiers so that the resulting formula is satisfied by m; if m |/= phi, then return None.