Module type Pa__PredicateAutomata.S

type t

type predicate

type letter

type letter_set

type formula = (predicate, int) Pa.PaFormula.formula

type config

module Config : Pa.Struct.S with type predicate = predicate and type t = config

module LetterSet : sig ... end

val pp : Stdlib.Format.formatter -> t -> unit

val pp_letter : Stdlib.Format.formatter -> letter -> unit

val pp_ground : int -> Stdlib.Format.formatter -> t -> unit

pp_ground size fmt a formats the alternating finite automata obtained from the predicate automaton a by instantiating the universe to size. The printed format is the input format of the Alaska tool.

val pp_formula : Stdlib.Format.formatter -> formula -> unit

val make : letter_set -> (predicate * int) list -> formula -> predicate list -> t

make alphabet vocabulary initial accepting creates a new PA with alphabet alphabet, vocabulary vocabulary, initial formula initial, and accepting predicates accept. Initially, the transition rule is defined to be delta(q,a) = false for all (q,a). Transitions can be added using add_transition.

val add_predicate : t -> predicate -> int -> unit

add_predicate pa p arity destructively updates the PA pa by adding the predicate p to its vocabulary, with arity arity

val add_accepting : t -> predicate -> unit

add_accepting pa p destructive updates the PA pa by making p an accepting predicate symbol

val add_transition : t -> predicate -> letter_set -> formula -> unit

add_transition pa q letters phi destructively updates the PA pa by adding (for each letter in letters) the transition rule delta(q,letter) = phi \/ psi, where psi was the previous transition rule for (q,letter).

val conjoin_transition : t -> predicate -> letter_set -> formula -> unit

conjoin_transition pa q letters phi destructively updates the PA pa by adding (for each letter in letters) the transition rule delta(q,letter) = phi /\ psi, where psi was the previous transition rule for (q,letter).

val alphabet : t -> letter_set

val vocabulary : t -> (predicate * int) BatEnum.t

vocabulary pa returns an enumeration of the predicate symbols used in pa along with their arity.

val mem_vocabulary : t -> predicate -> bool

mem_vocabularity pa p returns true iff p belongs to the vocabulary of pa.

val arity : t -> predicate -> int

arity pa p returns the arity of the predicate p. If p does not belong to the vocabulary of pa, raise Not_found.

val initial : t -> formula

val post : t -> formula -> letter -> formula

Compute the strongest post-condition of a formula under the transition relation corresponding to a given letter.

val concrete_post : t -> formula -> (letter * int) -> formula

Compute the strongest post-condition of a formula under the transition relation corresponding to a given indexed letter.

val succs : t -> config -> (letter * int) -> config BatEnum.t

val successors : t -> config -> int -> (letter * config) BatEnum.t

successors pa config index computes a set of index-labeled successors of a configuration config.

val pred : t -> config -> (letter * int) -> config

val accepting_formula : t -> formula -> bool

accepting_formula pa phi determines whether all models of phi are accepting configurations of pa.

val accepting : t -> config -> bool

val negate : t -> t

val union : t -> t -> t

val intersect : t -> t -> t